azcosmos: split GetChangeFeed into its own file + restore options test coverage
- Move GetChangeFeed, buildChangeFeedInitialQueue, resolveFeedRangeToChildren,
  buildChildQueueEntries, getChangeFeedForQueue, and serializeCompositeContinuationToken
  out of cosmos_container.go into cosmos_container_change_feed.go (mirrors the
  cosmos_container_read_many.go split). Pure relocation, no behavior change.
- Add cosmos_change_feed_request_options_test.go covering buildRequestHeaders:
  defaults (only AIM header), all-fields population, empty-ETag continuation
  is omitted, and PartitionKey serialization errors are surfaced (the new
  error-returning contract the legacy toHeaders silently swallowed).
